Tried a mroe data orienteated approach
Created a mother class to hold everything in it and a minimalistic scene... Added a clock to it too, several... Not much performance difference by numbers but noticed jumping in the images. I wonder if I can restrict my computer's performance for this, or if Unity is doing things a little differently on the back end... Either way no difference after 5k objects.

If only the rendering portion of the objects could be referenced as a reference. Or if it's actually in the GPU Cashe or memory which is why I can't directly modify it... Odd though.