DataDriven Attempt 2-4
Brought the clock down to 0.02ms on my rig from 2.2 ms. Process is showin in DataDriven2.cs through DataDriven4.cs.
The process on the OO style grew over time and took 2.3 or more ms per frame. This runs much lighter on the bus and allows for more things to happen in the meantime.

Everything was moved to maximize the CPU Cache and try to limit bus as much as possible.